requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
litellm>=1.6.0
python-dateutil>=2.8.2